% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_phenotype.R
\name{adjustphenotypes}
\alias{adjustphenotypes}
\title{Adjust Phenotypes by Genotype Mean}
\usage{
adjustphenotypes(genovec, phenovec)
}
\arguments{
\item{genovec}{Character vector of genotype/variety identifiers}

\item{phenovec}{Numeric vector of unadjusted phenotypic records, of the same
length as \code{genovec}.}
}
\value{
A list with two elements:
In position one, \code{adj_pheno} is a vector of adjusted phenotypes, ie
each individual's deviation from the mean of all individuals of the
same genotype.
In position two, \code{replicated} is a logical vector indicating whether
a given individual is from a genotype/variety that was replicated in the
field.
}
\description{
Description
}
