% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCARF.R
\name{SCARF}
\alias{SCARF}
\title{Perform Spatial Correction}
\usage{
SCARF(
  data,
  genoname,
  phenoname,
  predictornames,
  regressoroptions = list(ntree = 10),
  randomseed = NA
)
}
\arguments{
\item{data}{Data Frame or Tibble (?) containing all the data to be used by
SCARF}

\item{genoname}{String specifying the name of the column in \code{data} that
contains genotype/variety identifiers.}

\item{phenoname}{String specifying the name of the column in \code{data} that
contains phenotypic measurements. This column should be numeric.}

\item{predictornames}{String vector containing names of the columns in
\code{data} that are to be used as predictors by the Random Forest model
that models spatial variation. This will usually be positional coordinates,
such as Range and Row numbers.}

\item{regressoroptions}{List of named parameters to pass to
\code{randomForest::randomForest()}. Default is \code{list(ntree=10)}.}

\item{randomseed}{Integer to pass to \code{set.seed()}. Otherwise, results
can vary due to randomness in the training of the Random Forest model.}
}
\value{
List of length two:
In position 1, \code{results} has a data.frame
with columns corresponding to \code{genoname}, \code{phenoname},
\code{predictornames}, as well as the predicted \code{spatial_effect} for
each plot, and the adjusted phenotypes (raw phenotype - spatial correction).
In position 2, \code{model} has the Random Forest model that was fit to
the data and used to predict spatial effects.
}
\description{
Description
SCARF assumes data are from single field.
}
